Wait for events in the queue (#24)
# Why

Feedback from our users - in the server app it's fine to just wait a
second for events to send if the application doesn't end with an
unhandled exception. This change allows the setup a timeout for the
queue to clean up on the application exit. Thanks to this we can see all
reports from our demo, or we're using if someone is testing us with a
simple script, the data is in Backtrace

---------

Co-authored-by: Sebastian Alex <sebapotok@gmail.com>